PRO-407: update docs and API spec
- Add eslint config.
  - Ignore the build dir, ignore import errors stemming from docusaurus webpack aliases, and always include jsx files.
- Run eslint --fix across the repository.
  - Update file extensions of files that include JSX to .jsx.
- Add a top level enumeration of error codes and alphabetize error code sections.
- Add support for mermaid diagrams.
- Merge Node Types and Nodes sections in API spec.
- Update distribution reference and API spec to document the nuances of the `next_distribution_id` field.
- Update node groups reference to link to distributions reference.
