+++ src/docs/reference/node-groups.md
@@ -6,4 +6,4 @@ title: Node Groups
   <title>Ref | Node Groups</title>
 </head>
 
-Nodes groups are collections of nodes of a single node type. They are associated with a set of distributions. A node group's distributions form an anti-arborescence or in-tree.
+Nodes groups are collections of nodes of a single node type. They are associated with a set of [distributions](distributions). A node group's distributions form an anti-arborescence or in-tree.
